Refactor AG10 daemon and send surfaces
